UI improvements
Made it possible for the <> buttons to fade out after not moving the
mouse for a while. This allows clicking through a group while having the
buttons fade out if the mouse isn't moved, it's a little bit cleaner and
also works with keyboard navigation. Keeping the mouse on an actual
button instead of the larger click area will keep the buttons visible.

Also fixed a bug that could cause a clickable button to not have
cursor:pointer in some situations, making it seem unclickable.