bugfixes and added active states
fixed a bug that made it impossible to re-open Strip as it was closing
at the last couple of miliseconds.

elements with class='strip' will now get the 'strip-active-element'
class when opened, the entire group also gets the 'strip-active-group'
class.

fixed the broken image icon in IE7-8

the base and reset skins are no longer exposed