Workaround a visual glitch on videos in Chrome
Spam clicking an element that toggles a video could cause a visual
glitch in Chrome. It doesn't seem to handle an iframe loading in content
as it's being moved off screen before being removed. A visible element
is kept on the screen that isn't accessible through the DOM.

We work around this by instantly hiding the video just before Strip
hides, avoiding movement of the iframe as it is removed, and do this for
all browsers just in case.