Reverted the Chrome visual glitch workaround
1.1.0.2 didn't fix the glitch but allowed it to appear in the middle of
the screen, reverting because it was better on the edge of the screen.

The video/iframe glitch in Chrome when quickly toggling is a browser bug
that isn't as easily fixed, since the glitching element isn't accessible
through the DOM.

Further investigation is needed to find out what's causing this, and to
see if a proper workaround can be found.