Switched to a CSS workaround for the Chrome glitch
This fixes the visual glitch in Chrome when toggling a video quickly,
without affecting the entire page like the javascript workaround did.

We used to set "transform: translateZ(0px)" on the entire page in all
WebKit browsers for a short duration, which could cause other sorts of
conflicts. The new CSS workaround only affects the 'strp-window'
element.